requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.2.1
readability-lxml>=0.8.1
markdownify>=0.12.6
feedparser>=6.0.11
trafilatura>=1.12.0
tldextract>=5.1.2
charset-normalizer>=3.3.2
Flask>=3.0.0
requests-html==0.10.0